Added token matcher and layed out functions for parsers
