--- examples/JSON Parser/test.json
@@ -1,5 +1,0 @@
-{
-    "name": "Ryan",
-    "age": 21,
-    "height": 6.1
-}
