+++ examples/demo.rs
@@ -2,7 +2,7 @@ use flux_bnf::{bnf, lexer::CullStrategy};
 
 fn main() {
     let bnf_input = include_str!("../src/tests/bnf/json.bnf");
-    let test_input = include_str!("JSON Parser/test.json");
+    let test_input = include_str!("test.json");
 
     let mut lexer = match bnf::parse(bnf_input) {
         Ok(v) => v,
@@ -14,7 +14,7 @@ fn main() {
     };
 
     lexer.add_rule_for_names(
-        vec!["sep", "null"],
+        vec!["sep".to_string(), "null".to_string()],
         CullStrategy::LiftChildren,
     );
 
@@ -37,3 +37,4 @@ fn main() {
         .for_each(|t| println!("{:?}  {}", t.get_name(), t.get_match()));
     println!("{:#?}", root_token.first());
 }
+
