+++ examples/json.bnf
@@ -1,11 +1,11 @@
 root ::= object
 sep ::= [ \t\n]*
-object ::= string | integer | boolean | list | map | null
+object ::= boolean | null | string | decimal | integer | map | list
 string ::= "\"" ("\\" [^] | [^"])+ "\""
 integer ::= "-"? [0-9]+
 decimal ::= "-"? [0-9]+ "." [0-9]+
 boolean ::= "true" | "false"
-list ::= sep  "[" sep (object sep ("," sep object)*)? sep "]" sep
+list ::= sep "[" sep (object sep ("," sep object)*)? sep "]" sep
 map ::= "{" sep (mapEntry sep ("," sep mapEntry)*)? sep "}" sep
 mapEntry ::= string sep ":" sep object
 null ::= "null"
