+++ examples/test.json
@@ -1,0 +1,3 @@
+{
+    "number": 1
+}
